Added changes for window

+++ frontend/src/app/_services/storage.service.ts
@@ -1,6 +1,9 @@
 import { HttpClient } from "@angular/common/http";
-import { Injectable } from "@angular/core";
+import { Inject, Injectable } from "@angular/core";
 import { Router } from "@angular/router";
+
+import { PLATFORM_ID } from "@angular/core";
+import { isPlatformBrowser } from "@angular/common";
 
 const USER_KEY = "auth-user";
 const TOKEN = "token";
@@ -9,42 +12,57 @@ const TOKEN = "token";
   providedIn: "root"
 })
 export class StorageService {
-  constructor(private http: HttpClient, public router: Router) {}
+  constructor(@Inject(PLATFORM_ID) private platformId: any, private http: HttpClient, public router: Router) {}
 
   clean(): void {
-    window.sessionStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.clear();
+    }
   }
 
   public saveUser(user: any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.removeItem(USER_KEY);
+      window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    }
   }
 
   public getUser(): any {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return JSON.parse(user);
+    if (isPlatformBrowser(this.platformId)) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      if (user) {
+        return JSON.parse(user);
+      }
     }
-
     return null;
   }
 
   public isLoggedIn(): boolean {
-    let authToken = window.sessionStorage.getItem(TOKEN);
-    return authToken !== null ? true : false;
+    if (isPlatformBrowser(this.platformId)) {
+      let authToken = window.sessionStorage.getItem(TOKEN);
+      return authToken !== null ? true : false;
+    } else {
+      return false;
+    }
   }
   public getToken(): any {
-    return window.sessionStorage.getItem(TOKEN);
+    if (isPlatformBrowser(this.platformId)) {
+      return window.sessionStorage.getItem(TOKEN);
+    }
   }
 
   public setSession(token: any) {
-    window.sessionStorage.setItem(TOKEN, token);
+    if (isPlatformBrowser(this.platformId)) {
+      window.sessionStorage.setItem(TOKEN, token);
+    }
   }
 
   public doLogout() {
-    let removeToken = window.sessionStorage.removeItem(TOKEN);
-    if (removeToken == null) {
-      this.router.navigate(["login"]);
+    if (isPlatformBrowser(this.platformId)) {
+      let removeToken = window.sessionStorage.removeItem(TOKEN);
+      if (removeToken == null) {
+        this.router.navigate(["login"]);
+      }
     }
   }
 }
